requirements.txt
# Add Python dependencies here, e.g.
# fastapi==0.111.0
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
python-telegram-bot>=20.0
pyzbar>=0.1.9
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
